#72 - 댓/답글 관련 잇슈
- 기존 파일 대채 후 삭제
- 행 간격 사이 조절
