+++ IMAD_Project.xcodeproj/xcuserdata/quarang.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -796,8 +796,8 @@
             filePath = "IMAD_Project/Core/Comment/View/CommentDetailsView.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "53"
-            endingLineNumber = "53"
+            startingLineNumber = "56"
+            endingLineNumber = "56"
             landmarkName = "body"
             landmarkType = "24">
          </BreakpointContent>
